pyproject.toml
[project]
name = "bicomplex-pytorch"
version = "0.1.0"
description = "Neural networks with bicomplex numbers for PyTorch using idempotent representation"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "torch>=2.4.1",
    "numpy>=1.24.4",
    "pytest>=8.4.2",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}
keywords = ["deep-learning", "pytorch", "bicomplex", "neural-networks", "hypercomplex", "idempotent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/yourusername/bicomplex-pytorch"
Documentation = "https://bicomplex-pytorch.readthedocs.io"
Repository = "https://github.com/yourusername/bicomplex-pytorch"
"Bug Tracker" = "https://github.com/yourusername/bicomplex-pytorch/issues"
Changelog = "https://github.com/yourusername/bicomplex-pytorch/blob/main/CHANGELOG.md"

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=2.5.0",
    "pytest-benchmark>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=0.950",
    "pre-commit>=2.20.0",
]

# Documentation dependencies
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-autodoc-typehints>=1.18.0",
    "myst-parser>=0.18.0",
    "sphinxcontrib-napoleon>=0.7",
]

# Example dependencies
examples = [
    "matplotlib>=3.5.0",
    "jupyter>=1.0.0",
    "jupyterlab>=3.0.0",
    "seaborn>=0.11.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
]

# Benchmarking dependencies
benchmark = [
    "pytest-benchmark>=4.0.0",
    "memory-profiler>=0.60.0",
    "line-profiler>=4.0.0",
]

# All optional dependencies
all = [
    "bicomplex-pytorch[dev,docs,examples,benchmark]",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
version = {source = "file", path = "bicomplex_pytorch/__version__.py"}
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=2.5.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=0.950",
    "pre-commit>=2.20.0",
]

[tool.pdm.scripts]
# Testing scripts
test = "pytest tests/ -v"
test-cov = "pytest tests/ -v --cov=bicomplex_pytorch --cov-report=html --cov-report=term"
test-fast = "pytest tests/ -v -x"  # Stop on first failure
test-parallel = "pytest tests/ -v -n auto"  # Run tests in parallel

# Code quality scripts
format = "black bicomplex_pytorch/ tests/ examples/"
lint = "ruff check bicomplex_pytorch/ tests/ examples/"
type-check = "mypy bicomplex_pytorch/ --ignore-missing-imports"
check-all = {composite = ["format", "lint", "type-check"]}

# Documentation scripts
docs-build = "sphinx-build -b html docs/ docs/_build"
docs-serve = "python -m http.server --directory docs/_build 8000"
docs-clean = "rm -rf docs/_build"

# Example scripts
run-basic = "python examples/basic_usage.py"
run-mnist = "python examples/mnist_classifier.py"

# Benchmarking scripts
benchmark = "pytest benchmarks/ --benchmark-only"
profile = "python -m cProfile -o profile.stats examples/basic_usage.py"

# Pre-commit setup
pre-commit-install = "pre-commit install"
pre-commit-run = "pre-commit run --all-files"

[tool.pdm.build]
includes = ["bicomplex_pytorch/"]
excludes = ["tests/", "docs/", "examples/", "benchmarks/"]

# Black configuration
[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff configuration (faster alternative to flake8, isort, etc.)
[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.isort]
known-first-party = ["bicomplex_pytorch"]

# MyPy configuration
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "complexPyTorch.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=bicomplex_pytorch",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests as requiring GPU (deselect with '-m \"not gpu\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["bicomplex_pytorch"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.pdm.resolution]
respect-source-order = true

[tool.pdm.resolution.overrides]
numpy = ">=2.0.0"
torch = ">=2.7.0"
setuptools = ">=80.0.0"
typing-extensions = ">=4.15.0"
filelock = ">=3.20.0"
fsspec = ">=2025.10.0"
markupsafe = ">=3.0.0"
networkx = ">=3.5"
sympy = ">=1.14.0"